huhhh

+++ tests/bpf_shim.h
@@ -15,7 +15,7 @@ static void *(* const bpf_map_lookup_elem)(void *map, const void *key) = (void *
 static long (* const bpf_map_update_elem)(void *map, const void *key, const void *value, unsigned long long flags) = (void *) 2;
 static long (* const bpf_map_delete_elem)(void *map, const void *key) = (void *) 3;
 static unsigned int (* const bpf_iter_num_new)(struct bpf_iter_num *it, int start, int end) = (void *) 5;
-static int *(* const bpf_iter_num_next)(struct bpf_iter_num *it) = (void *) 6;
+static long *(* const bpf_iter_num_next)(struct bpf_iter_num *it) = (void *) 6;
 static void (* const bpf_iter_num_destroy)(struct bpf_iter_num *it) = (void *) 7;
 static int (* const bpf_rand_int)(int min, int max) = (void *) 8;
